refactor(#126): GlobalExceptionHandler 3중 복제 → BaseExceptionHandler 추출
- nextup-infrastructure에 BaseExceptionHandler 추상 클래스 생성
- 3개 API 모듈의 GlobalExceptionHandler가 Base를 상속
- 중복 에러 처리 로직 제거 (162줄 삭제, 93줄 추가)
- 모듈 간 에러 응답 일관성 보장
- scorer 모듈: handleBusinessException/handleException override로 기존 동작 보존
