+++ nextup-api/src/main/kotlin/com/nextup/api/exception/GlobalExceptionHandler.kt
@@ -2,22 +2,17 @@ package com.nextup.api.exception
 
 import com.nextup.common.dto.ApiResponse
 import com.nextup.common.exception.AuthenticationException
-import com.nextup.common.exception.BusinessException
 import com.nextup.common.exception.ForbiddenException
 import com.nextup.common.exception.InvalidStateException
-import com.nextup.common.exception.NotFoundException
-import org.slf4j.LoggerFactory
+import com.nextup.infrastructure.exception.BaseExceptionHandler
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.http.converter.HttpMessageNotReadableException
-import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.MissingServletRequestParameterException
 import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestControllerAdvice
 
 @RestControllerAdvice
-class GlobalExceptionHandler {
-    private val log = LoggerFactory.getLogger(javaClass)
+class GlobalExceptionHandler : BaseExceptionHandler() {
 
     @ExceptionHandler(AuthenticationException::class)
     fun handleAuthenticationException(ex: AuthenticationException): ResponseEntity<ApiResponse<Nothing>> {
@@ -35,38 +30,12 @@ class GlobalExceptionHandler {
             .body(ApiResponse.error(ex.code, ex.message))
     }
 
-    @ExceptionHandler(NotFoundException::class)
-    fun handleNotFoundException(ex: NotFoundException): ResponseEntity<ApiResponse<Nothing>> {
-        log.warn("NotFoundException: code={}, message={}", ex.code, ex.message)
-        return ResponseEntity
-            .status(HttpStatus.NOT_FOUND)
-            .body(ApiResponse.error(ex.code, ex.message))
-    }
-
     @ExceptionHandler(InvalidStateException::class)
     fun handleInvalidStateException(ex: InvalidStateException): ResponseEntity<ApiResponse<Nothing>> {
         log.warn("InvalidStateException: code={}, message={}", ex.code, ex.message)
         return ResponseEntity
             .status(HttpStatus.BAD_REQUEST)
             .body(ApiResponse.error(ex.code, ex.message))
-    }
-
-    @ExceptionHandler(BusinessException::class)
-    fun handleBusinessException(ex: BusinessException): ResponseEntity<ApiResponse<Nothing>> {
-        log.warn("BusinessException: code={}, message={}", ex.code, ex.message)
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(ApiResponse.error(ex.code, ex.message))
-    }
-
-    @ExceptionHandler(HttpMessageNotReadableException::class)
-    fun handleHttpMessageNotReadableException(
-        ex: HttpMessageNotReadableException,
-    ): ResponseEntity<ApiResponse<Nothing>> {
-        log.warn("HttpMessageNotReadableException: {}", ex.message)
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(ApiResponse.error("INVALID_REQUEST", "요청 본문을 읽을 수 없습니다"))
     }
 
     @ExceptionHandler(MissingServletRequestParameterException::class)
@@ -79,38 +48,11 @@ class GlobalExceptionHandler {
             .body(ApiResponse.error("MISSING_PARAMETER", ex.message))
     }
 
-    @ExceptionHandler(MethodArgumentNotValidException::class)
-    fun handleValidationException(ex: MethodArgumentNotValidException): ResponseEntity<ApiResponse<Nothing>> {
-        val errors =
-            ex.bindingResult.fieldErrors
-                .joinToString("; ") { "${it.field}: ${it.defaultMessage}" }
-        log.warn("ValidationException: {}", errors)
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(ApiResponse.error("VALIDATION_ERROR", errors))
-    }
-
-    @ExceptionHandler(IllegalArgumentException::class)
-    fun handleIllegalArgumentException(ex: IllegalArgumentException): ResponseEntity<ApiResponse<Nothing>> {
-        log.warn("IllegalArgumentException: {}", ex.message)
-        return ResponseEntity
-            .status(HttpStatus.BAD_REQUEST)
-            .body(ApiResponse.error("INVALID_ARGUMENT", ex.message ?: "Invalid argument"))
-    }
-
     @ExceptionHandler(IllegalStateException::class)
     fun handleIllegalStateException(ex: IllegalStateException): ResponseEntity<ApiResponse<Nothing>> {
         log.warn("IllegalStateException: {}", ex.message)
         return ResponseEntity
             .status(HttpStatus.BAD_REQUEST)
             .body(ApiResponse.error("INVALID_STATE", ex.message ?: "Invalid state"))
-    }
-
-    @ExceptionHandler(Exception::class)
-    fun handleException(ex: Exception): ResponseEntity<ApiResponse<Nothing>> {
-        log.error("UnexpectedException: {}", ex.message, ex)
-        return ResponseEntity
-            .status(HttpStatus.INTERNAL_SERVER_ERROR)
-            .body(ApiResponse.error("INTERNAL_ERROR", "An unexpected error occurred"))
     }
 }
